+++ wp-parsely/src/Endpoints/user-meta/class-base-endpoint-user-meta.php
@@ -199,26 +199,43 @@ abstract class Base_Endpoint_User_Meta extends Base_Endpoint {
 	 * @since 3.14.0 Added support for nested arrays.
 	 *
 	 * @param array<string, mixed> $meta_value The meta value to sanitize.
+	 * @param string               $parent_key  The parent key for the current level of the meta.
 	 * @return array<string, mixed> The sanitized meta as an array of subvalues.
 	 */
-	protected function sanitize_value( array $meta_value ): array {
+	protected function sanitize_value( array $meta_value, string $parent_key = '' ): array {
 		$sanitized_value = array();
 
-		foreach ( $meta_value as $key => $value ) {
-			$key = sanitize_text_field( $key );
+		// Determine the current level's specifications based on the parent key.
+		/**
+		 * Current level's specifications.
+		 *
+		 * @var array<string, Subvalue_Spec> $current_specs
+		 */
+		$current_specs = ( '' === $parent_key ) ? $this->get_subvalues_specs() : $this->get_nested_specs( $parent_key );
 
-			// Skip if the key isn't valid.
-			if ( ! array_key_exists( $key, $this->valid_subvalues ) ) {
-				continue;
+		foreach ( $current_specs as $key => $spec ) {
+			$composite_key = '' === $parent_key ? $key : $parent_key . '.' . $key;
+
+			// Check if the key exists in the input meta value array.
+			if ( array_key_exists( $key, $meta_value ) ) {
+				$value = $meta_value[ $key ];
+			} else {
+				// Key is missing in the input, use the default value from the specifications.
+				$value = $this->get_default( explode( '.', $composite_key ) );
 			}
 
-			// Use the enhanced sanitize_subvalue method.
-			$sanitized_value[ $key ] = $this->sanitize_subvalue( $key, $value );
-		}
-
-		// If not all subvalues are set, return the default meta value.
-		if ( count( array_diff_key( $this->valid_subvalues, $sanitized_value ) ) !== 0 ) {
-			return $this->default_value;
+			/**
+			 * Spec for the current key.
+			 *
+			 * @var array{default: mixed, values?: array<mixed>} $spec
+			 */
+			if ( is_array( $value ) && isset( $spec['values'] ) ) {
+				// Recursively handle nested arrays if 'values' spec exists for this key.
+				$sanitized_value[ $key ] = $this->sanitize_value( $value, $composite_key );
+			} else {
+				// Directly sanitize non-array values or non-nested specs.
+				$sanitized_value[ $key ] = $this->sanitize_subvalue( $composite_key, $value );
+			}
 		}
 
 		return $sanitized_value;
@@ -227,41 +244,146 @@ abstract class Base_Endpoint_User_Meta extends Base_Endpoint {
 	/**
 	 * Sanitizes the passed subvalue.
 	 *
-	 * @since 3.13.0
 	 * @since 3.14.0 Added support for nested arrays.
+	 * @since 3.13.0
 	 *
-	 * @param string $key The subvalue's key.
+	 * @param string $composite_key The subvalue's key.
 	 * @param mixed  $value The value to sanitize.
 	 * @return mixed The sanitized subvalue.
 	 */
-	protected function sanitize_subvalue( string $key, $value ) {
-		// Handle nested arrays recursively.
+	protected function sanitize_subvalue( string $composite_key, $value ) {
+		$keys         = explode( '.', $composite_key );
+		$valid_values = $this->get_valid_values( $keys );
+
 		if ( is_array( $value ) ) {
-			$sanitized_array = array();
-			foreach ( $value as $subkey => $subvalue ) {
-				// Sanitize keys of nested arrays.
-				$sanitized_subkey = sanitize_text_field( $subkey );
-				// Recursively sanitize each value in the nested array.
-				$sanitized_array[ $sanitized_subkey ] = $this->sanitize_subvalue( $sanitized_subkey, $subvalue );
+			// Check if $value elements are inside $valid_values
+			// If not, the value should be the default value.
+			$valid_value = array();
+			foreach ( $value as $key => $val ) {
+				if ( in_array( $val, $valid_values, true ) ) {
+					$valid_value[ $key ] = $val;
+				}
 			}
-			return $sanitized_array;
+			return $valid_value;
 		}
 
-		// Sanitize simple values.
 		if ( is_string( $value ) ) {
 			$value = sanitize_text_field( $value );
 		}
 
-		// Allow any value when no valid subvalues are given.
-		if ( ! isset( $this->valid_subvalues[ $key ] ) || count( $this->valid_subvalues[ $key ] ) === 0 ) {
+		if ( count( $valid_values ) === 0 ) {
 			return $value;
 		}
 
-		// Use default value if the actual value is not valid.
-		if ( ! in_array( $value, $this->valid_subvalues[ $key ], true ) ) {
-			$value = $this->default_value[ $key ];
+		if ( ! in_array( $value, $valid_values, true ) ) {
+			return $this->get_default( $keys );
 		}
 
 		return $value;
+	}
+
+	/**
+	 * Checks if a given composite key is valid.
+	 *
+	 * @since 3.14.3
+	 *
+	 * @param string|mixed $composite_key The composite key representing the nested path.
+	 * @return bool Whether the key is valid.
+	 */
+	protected function is_valid_key( $composite_key ): bool {
+		if ( ! is_string( $composite_key ) ) {
+			return false; // Key path is not a string.
+		}
+
+		$keys    = explode( '.', $composite_key );
+		$current = $this->valid_subvalues;
+
+		foreach ( $keys as $key ) {
+			if ( ! is_array( $current ) || ! isset( $current[ $key ] ) ) {
+				return false; // Key path is invalid.
+			}
+
+			if ( isset( $current[ $key ]['values'] ) ) {
+				$current = $current[ $key ]['values'];
+			} else {
+				$current = $current[ $key ];
+			}
+		}
+
+		return true;
+	}
+
+	/**
+	 * Gets the valid values for a given setting path.
+	 *
+	 * @since 3.14.3
+	 *
+	 * @param array<string> $keys The path to the setting.
+	 * @return array<mixed> The valid values for the setting path.
+	 */
+	protected function get_valid_values( array $keys ): array {
+		$current = $this->valid_subvalues;
+
+		foreach ( $keys as $key ) {
+			if ( ! is_array( $current ) || ! isset( $current[ $key ] ) ) {
+				return array(); // No valid values for invalid key path.
+			}
+			if ( isset( $current[ $key ]['values'] ) ) {
+				$current = $current[ $key ]['values'];
+			} else {
+				$current = $current[ $key ];
+			}
+		}
+
+		return is_array( $current ) ? $current : array();
+	}
+
+	/**
+	 * Gets the default value for a given setting path.
+	 *
+	 * @since 3.14.3
+	 *
+	 * @param array<string> $keys The path to the setting.
+	 * @return mixed|array<mixed>|null The default value for the setting path.
+	 */
+	protected function get_default( array $keys ) {
+		$current = $this->default_value;
+
+		foreach ( $keys as $key ) {
+			if ( ! is_array( $current ) || ! isset( $current[ $key ] ) ) {
+				return null; // No default value for invalid key path.
+			}
+			if ( isset( $current[ $key ]['default'] ) ) {
+				$current = $current[ $key ]['default'];
+			} else {
+				$current = $current[ $key ];
+			}
+		}
+
+		return $current; // Return default value for valid key path.
+	}
+
+
+	/**
+	 * Gets the specifications for nested settings based on a composite key.
+	 *
+	 * @since 3.14.3
+	 *
+	 * @param string $composite_key The composite key representing the nested path.
+	 * @return array<mixed> The specifications for the nested path.
+	 */
+	protected function get_nested_specs( string $composite_key ): array {
+		$keys  = explode( '.', $composite_key );
+		$specs = $this->get_subvalues_specs();
+
+		foreach ( $keys as $key ) {
+			if ( is_array( $specs[ $key ] ) && array_key_exists( 'values', $specs[ $key ] ) ) {
+				$specs = $specs[ $key ]['values'];
+			} else {
+				break;
+			}
+		}
+
+		return $specs;
 	}
 }