+++ wp-parsely/src/content-helper/common/settings/types/index.ts
@@ -2,9 +2,11 @@
  * Import the settings types.
  */
 import type {
-	SidebarSettings,
-	TitleSuggestionsSettings,
 	PerformanceStatsSettings,
+	RelatedPostsSettings,
+	SidebarSettings,
+	SmartLinkingSettings,
+	TitleSuggestionsSettings,
 } from './sidebar-settings';
 import type { TopPostsSettings } from './top-posts-settings';
 
@@ -12,9 +14,11 @@ import type { TopPostsSettings } from './top-posts-settings';
  * Export the settings types.
  */
 export type {
-	SidebarSettings,
-	TitleSuggestionsSettings, // Part of SidebarSettings type.
 	PerformanceStatsSettings, // Part of SidebarSettings type.
+	RelatedPostsSettings, // Part of SidebarSettings type.
+	SidebarSettings,
+	SmartLinkingSettings, // Part of SidebarSettings type.
+	TitleSuggestionsSettings, // Part of SidebarSettings type.
 	TopPostsSettings,
 };