+++ wp-parsely/src/content-helper/editor-sidebar/performance-stats/component.tsx
@@ -32,8 +32,8 @@ import { PerformanceCategoriesPanel } from './component-panel-categories';
 import { PerformanceOverviewPanel } from './component-panel-overview';
 import { PerformanceReferrersPanel } from './component-panel-referrers';
 import { PerformanceData } from './model';
-import './performance-details.scss';
-import { PerformanceDetailsProvider } from './provider';
+import './performance-stats.scss';
+import { PerformanceStatsProvider } from './provider';
 
 // Number of attempts to fetch the data before displaying an error.
 const FETCH_RETRIES = 1;
@@ -80,7 +80,7 @@ const availablePanels: PanelDescriptor[] = [
  * @return { boolean } True if the panel is visible, false otherwise.
  */
 const isPanelVisible = ( settings: SidebarSettings, panel: string ): boolean => {
-	return settings.PerformanceStatsSettings.VisiblePanels.includes( panel );
+	return settings.PerformanceStats.VisiblePanels.includes( panel );
 };
 
 /**
@@ -111,13 +111,13 @@ const PerformanceStatsMenu = (
 			return;
 		}
 
-		// Check if the panel is in the settings.PerformanceStatsSettings.VisiblePanels array
+		// Check if the panel is in the settings.PerformanceStats.VisiblePanels array
 		// If it is, remove it with setSettings, if not, add it.
 		if ( isPanelVisible( settings, panel ) ) {
 			setSettings( {
-				PerformanceStatsSettings: {
-					...settings.PerformanceStatsSettings,
-					VisiblePanels: settings.PerformanceStatsSettings.VisiblePanels.filter(
+				PerformanceStats: {
+					...settings.PerformanceStats,
+					VisiblePanels: settings.PerformanceStats.VisiblePanels.filter(
 						( p ) => p !== panel
 					),
 				},
@@ -125,9 +125,9 @@ const PerformanceStatsMenu = (
 			Telemetry.trackEvent( 'editor_sidebar_performance_panel_closed', { panel } );
 		} else {
 			setSettings( {
-				PerformanceStatsSettings: {
-					...settings.PerformanceStatsSettings,
-					VisiblePanels: [ ...settings.PerformanceStatsSettings.VisiblePanels, panel ],
+				PerformanceStats: {
+					...settings.PerformanceStats,
+					VisiblePanels: [ ...settings.PerformanceStats.VisiblePanels, panel ],
 				},
 			} );
 			Telemetry.trackEvent( 'editor_sidebar_performance_panel_opened', { panel } );
@@ -153,8 +153,8 @@ const PerformanceStatsMenu = (
 	 */
 	const resetAll = (): void => {
 		setSettings( {
-			PerformanceStatsSettings: {
-				...settings.PerformanceStatsSettings,
+			PerformanceStats: {
+				...settings.PerformanceStats,
 				VisiblePanels: availablePanels.map( ( panel ) => panel.name ),
 			},
 		} );
@@ -211,10 +211,10 @@ export const PerformanceStats = (
 	const { settings, setSettings } = useSettings<SidebarSettings>();
 
 	useEffect( () => {
-		const provider = new PerformanceDetailsProvider();
+		const provider = new PerformanceStatsProvider();
 
 		const fetchPosts = async ( retries: number ) => {
-			provider.getPerformanceDetails( period )
+			provider.getPerformanceStats( period )
 				.then( ( result ) => {
 					setPostDetails( result );
 					setLoading( false );
@@ -248,7 +248,7 @@ export const PerformanceStats = (
 				<div className="panel-settings">
 					<SelectControl
 						size="__unstable-large"
-						value={ settings.PerformanceStatsSettings.Period }
+						value={ settings.PerformanceStats.Period }
 						prefix={
 							<InputControlPrefixWrapper>
 								{ __( 'Period: ', 'wp-parsely' ) }
@@ -257,8 +257,8 @@ export const PerformanceStats = (
 						onChange={ ( selection ) => {
 							if ( isInEnum( selection, Period ) ) {
 								setSettings( {
-									PerformanceStatsSettings: {
-										...settings.PerformanceStatsSettings,
+									PerformanceStats: {
+										...settings.PerformanceStats,
 										Period: selection as Period,
 									},
 								} );